create function that gets and analyzes
